requirements.txt
selenium
aiohttp
webdriver-manager
psutil
beautifulsoup4
requests
streamlit
transformers
torch
torchvision
Pillow
scikit-image
fastapi==0.115.0
uvicorn==0.30.6
python-multipart
urllib3
motor==3.6.0
pymongo==4.9.1
pydantic
python-dotenv
cloudinary
opencv-contrib-python==4.10.0.84  # Replaced opencv-python-headless
openai